Post-synchronization fixing possible reason for bug on controverse website
some svg files didn't appear and one pdf didn't open. We think it's
because some had an accent in the title (Légende%20gauche.svg and
Texte%20arbre%20des%20débats.svg and the other one was named
Acteurs_Poles.svg in the code but Acteurs_poles.svg in the folder.  It
worked on our computers just fine but we suppose these are the reasons
why they didn't work on the website. We renamed the files with the
accents and wrote Acteurs_poles.svg in the code (instead of the same
with a capital P)